agents/pyproject.toml
[project]
name = "lightning-bounty-agents"
version = "0.1.0"
description = "Reference agents for the Lightning Bounty Marketplace"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "openai>=1.54.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "shared"]
